Create "edit this page" footer
This was a feature previously removed as the theme was undergoing
transitions. It has been re-added to the footer. Links are based
on the config options:
- repoHost
- repository
- docFolder
which allows for specifying custom repository targets (most notably
to use gitlab editing)

Most links are currently not working properly as the target master
branch still has the old filing logic